# api/app/swagger.yml
swagger: "2.0"
info:
  title: Movie Database
  description: This API provides endpoints for accessing a movies and actors database.
  version: 0.0.2
securityDefinitions:
  basicAuth:
    type: basic
definitions:
...
  User:
    type: object
    properties:
      login:
        type: string
      hashed_password:
        type: string
      role:
        type: string
  UserRequest:
    type: object
    properties:
      login:
        type: string
      password:
        type: string
responses:
  UnauthorizedError:
    description: Authentication information is missing or invalid
    headers:
      WWW_Authenticate:
        type: string
paths:
...
  /users:
    get:
      summary: Get all users
      security:
        - basicAuth: []
      responses:
        200:
          description: ok
          schema:
            type: array
            items:
              $ref: "#/definitions/User"
        401:
          $ref: '#/responses/UnauthorizedError'
        403:
          description: Forbidden
  /register:      
    post:
      summary: Register a new user
      consumes:
        - application/json
      parameters:
        - in: body
          name: user
          description: The user to add.
          schema:
            $ref: '#/definitions/UserRequest'
      responses:
        201:
          description: Created
        409:
          description: Conflict      